ENH: Create Task Extensions for Timeout Handling - Moves existing methods to be extensions - Updates previous references to use the extension syntax
